Move error handling to new middlewares dir
