+++ server/index.js
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
-/* eslint no-unused-vars: 0, no-underscore-dangle: 0, no-console: 0 */
+/* eslint no-console: 0 */
 
 const dotenv = require('dotenv');
 // load config from .env file
 const config = dotenv.config();
-const _ = require('lodash');
 const express = require('express');
 const bodyParser = require('body-parser');
+const errorHandling = require('../middlewares/error-handling.js');
 const authentication = require('./authentication.js');
 const getMediaApi = require('./get/index.js');
 
@@ -21,20 +21,7 @@ server.get('/', (req, res) => {
 
 server.get('/get', getMediaApi);
 
-server.use((err, req, res, next) => {
-  if (err.errors) {
-    const error = _.head(err.errors);
-    const status = _.head(err._headers.status);
-    const statusCode = status.substring(0, 3);
-    const statusText = status.substring(4, status.length);
-
-    console.error(`Twitter API error ${error.code}: ${error.message}`);
-    res.status(statusCode).send(statusText);
-  } else {
-    console.error(err.stack);
-    res.status(500).send(err.message);
-  }
-});
+server.use(errorHandling);
 
 function start() {
   server.listen(port, async () => {
